Move PersistenQueue to separate module
